Rename: Type Import edit
