+++ components/game/invitation/InvitationCard.tsx
@@ -1,23 +1,23 @@
-import * as Type from '@/lib/types';
-import {User} from '@/lib/classes/User';
+import {userStatusType, userInfoType} from '@/types/user';
+import {User} from '@/classes/User';
 import UserInfoCard from '@/components/card/userInfoCard/UserInfoCard';
 import {AccordionItem} from '@radix-ui/react-accordion';
 
-function userPropsToUserClass(props: InvitationCardProps): Type.UserInfo {
+function userPropsToUserClass(props: InvitationCardProps): userInfoType {
   const userClass = new User();
   userClass.id = props.id;
   userClass.profileImage = props.profileImage;
   userClass.name = props.name;
-  userClass.currentStatus = props.currentStatus as Type.userStatus;
+  userClass.currentStatus = props.currentStatus as userStatusType;
   userClass.introduction = props.introduction;
-  return userClass as Type.UserInfo;
+  return userClass as userInfoType;
 }
 
 interface InvitationCardProps {
   id: string;
   profileImage: string;
   name: string;
-  currentStatus: Type.userStatus | string;
+  currentStatus: userStatusType | string;
   introduction: string;
   isFriend: boolean;
   isBlocked: boolean;
